requirements.txt
django==3.2.7
djgentelella==0.2.7
djangorestframework==3.11.0
